Исправление функций по 11 заданию
