+++ templates/layout.php
@@ -51,7 +51,7 @@
         <ul class="nav__list container">
             <? foreach ($categories as $item): ?>
                 <li class="nav__item">
-                    <a href="all-lots.html"><?= $item; ?></a>
+                    <a href="all-lots.html"><?= $item['cat_name']; ?></a>
                 </li>
             <? endforeach; ?>
         </ul>
